+++ scripts/socketserver/echo.py
@@ -1,12 +1,13 @@
 #!/usr/bin/env python3
 #pylint:disable=missing-docstring,invalid-name
 
+import argparse
 import asyncio
 import websockets
 import sys
+import time
 
 HOST = "localhost"
-PORT = 4444
 
 def log(data):
     print(data, flush=True)
@@ -26,8 +27,23 @@ def connection_handler(connection, path):
         log("{} closed connection".format(connection_id))
 
 if __name__ == "__main__":
-    log("Starting server at {}:{}".format(HOST, PORT))
-    server = websockets.serve(connection_handler, HOST, PORT)
+    parser = argparse.ArgumentParser()
+    parser.add_argument("--port",
+        help="Port to listen on",
+        type=int,
+        default=4444)
+    parser.add_argument("--delay",
+        help="Time in seconds that a connection will be delayed before establishing it",
+        type=int,
+        default=0)
+    args = parser.parse_args()
+
+    def delaying_process_request(path, request_headers):
+        time.sleep(args.delay)
+        return None
+
+    log("Starting server at {}:{}".format(HOST, args.port))
+    server = websockets.serve(connection_handler, HOST, args.port, process_request=delaying_process_request)
     log("Running now.")
 
     asyncio.get_event_loop().run_until_complete(server)